Tcl-port of temper.py
(json output not yet implemented)

For now it is a mostly 1:1 translation, but I intend to make it
more "Tcl"ish: replace repeated code by a structure describing
supported venderid:productid pairs and firmware-strings with a
generic description of the peculiarities of each.

Another plan for future: Allow the caller to "force" the firmware,
just so it can be hardcoded in custom scripts, obviating the need
for the script to query it each time.  For people who have only
one type of thermometer device, anyway.